requirements.txt
pyspark==3.5.4
certifi==2024.6.2
python-dotenv==1.0.1
pinecone==5.4.2
pymongo==4.6.1
nltk==3.9.1
numpy==1.26.4
torch==2.2.2
scikit-learn==1.3.2
transformers==4.49.0
confluent-kafka>=2.3.0
python-dotenv
schedule==1.2.0